Fix TLS access in JIT with Musl
The factor 16 originates from 71cccc0bcc5f083ae6022face1c2514b5240b96a,
but it's probably a merge error of 94ba883e195613d5a01be07b6ca9a73024b5575b,
which used the right value.

The magic value is the size of a DTV element, which is a uintptr_t in musl.